Fix: Facturi-emise - colorare rânduri, spațiere badge, tipuri BigQuery
Rezolvate 3 probleme raportate:

1. Colorare rânduri nu funcționa (Tailwind JIT)
   - Înlocuit clase Tailwind dinamice cu inline styles
   - Verde: facturi încasate complet
   - Portocaliu: încasări parțiale

2. Lipsea spațiu între status și sumă ("Incasat18.150,00")
   - Restructurat badge-ul cu div-uri separate
   - Fiecare element pe linie proprie cu mb-1

3. Eroare BigQuery: FLOAT64 cannot be inserted into NUMERIC
   - Corectat tipurile în incasare/route.ts:
     - sumaIncasata: NUMERIC (pentru EtapeFacturi_v2)
     - suma_tranzactie, suma_target, etc: NUMERIC (TranzactiiMatching_v2)

+++ app/api/iapp/facturi-emise/incasare/route.ts
@@ -516,7 +516,7 @@ export async function POST(request: NextRequest) {
         types: {
           statusIncasare: 'STRING',
           dataIncasare: 'STRING',
-          sumaIncasata: 'FLOAT64',
+          sumaIncasata: 'NUMERIC',
           userId: 'STRING',
           facturaId: 'STRING'
         },
@@ -608,10 +608,10 @@ export async function POST(request: NextRequest) {
           tranzactieId: 'STRING',
           facturaId: 'STRING',
           targetDetails: 'STRING',
-          sumaIncasata: 'FLOAT64',
-          totalFactura: 'FLOAT64',
-          diferentaRon: 'FLOAT64',
-          diferentaProcent: 'FLOAT64',
+          sumaIncasata: 'NUMERIC',
+          totalFactura: 'NUMERIC',
+          diferentaRon: 'NUMERIC',
+          diferentaProcent: 'NUMERIC',
           matchingDetails: 'STRING',
           userId: 'STRING'
         },
